Fixed tile parallelism, tickspersec setting, build cleanup
