+++ source/Lib/CommonLib/Quant.cpp
@@ -129,12 +129,104 @@ QpParam::QpParam(const TransformUnit& tu, const ComponentID &compID, const bool 
 // ====================================================================================================================
 // Quant class member functions
 // ====================================================================================================================
-static void QuantCore(const CCoeffBuf&  piCoef,CoeffSigBuf piQCoef,TCoeff &uiAbsSum,TCoeff *deltaU,const int maxNumberOfCoeffs,const int defaultQuantisationCoefficient,const int iQBits,const int64_t iAdd,const TCoeff entropyCodingMinimum,const TCoeff entropyCodingMaximum ,const bool signHiding)
+static void QuantCore(const TransformUnit tu, const ComponentID compID, const CCoeffBuf& piCoef,CoeffSigBuf piQCoef,TCoeff &uiAbsSum, int &lastScanPos,TCoeff *deltaU,const int defaultQuantisationCoefficient,const int iQBits,const int64_t iAdd,const TCoeff entropyCodingMinimum,const TCoeff entropyCodingMaximum,const bool signHiding, const TCoeff m_thrVal)
 {
+  CoeffCodingContext cctx( tu, compID, signHiding );
+
+  const SPS &sps            = *tu.cs->sps;
+  const CompArea &rect      = tu.blocks[compID];
+  const uint32_t uiWidth    = rect.width;
+  const uint32_t uiHeight   = rect.height;
+  const ChannelType chType  = toChannelType( compID );
+  const int channelBitDepth = sps.bitDepths[ chType ];
+
+  const bool extendedPrecision     = sps.spsRExt.extendedPrecisionProcessing;
+  const int  maxLog2TrDynamicRange = sps.getMaxLog2TrDynamicRange(chType);
+
+  /* for 422 chroma blocks, the effective scaling applied during transformation is not a power of 2, hence it cannot be
+  * implemented as a bit-shift (the quantised result will be sqrt(2) * larger than required). Alternatively, adjust the
+  * uiLog2TrSize applied in iTransformShift, such that the result is 1/sqrt(2) the required result (i.e. smaller)
+  * Then a QP+3 (sqrt(2)) or QP-3 (1/sqrt(2)) method could be used to get the required result
+  */
+
+  // Represents scaling through forward transform
+  int iTransformShift = getTransformShift(channelBitDepth, rect.size(), maxLog2TrDynamicRange);
+
+  if (tu.mtsIdx[compID]==MTS_SKIP && extendedPrecision)
+  {
+    iTransformShift = std::max<int>(0, iTransformShift);
+  }
+
+  const uint32_t log2CGSize         = cctx.log2CGSize();
+
+  uiAbsSum = 0;
+
+  const int iCGSize   = 1 << log2CGSize;
+
+  const uint32_t lfnstIdx = tu.cu->lfnstIdx;
+  const int iCGNum   = lfnstIdx > 0 ? 1 : std::min<int>(JVET_C0024_ZERO_OUT_TH, uiWidth) * std::min<int>(JVET_C0024_ZERO_OUT_TH, uiHeight) >> cctx.log2CGSize();
+  int       iScanPos = ( iCGNum << log2CGSize ) - 1;
+
+  if( lfnstIdx > 0 && ( ( uiWidth == 4 && uiHeight == 4 ) || ( uiWidth == 8 && uiHeight == 8 ) ) )
+  {
+    iScanPos = 7;
+  }
+
+  // Find first non-zero coeff
+  for( ; iScanPos > 0; iScanPos-- )
+  {
+    uint32_t uiBlkPos = cctx.blockPos( iScanPos );
+    if( piCoef.buf[uiBlkPos] )
+      break;
+  }
+
+  //////////////////////////////////////////////////////////////////////////
+  //  Loop over sub-sets (coefficient groups)
+  //////////////////////////////////////////////////////////////////////////
+  
+  TCoeff thres = 0, useThres = 0;
+  
+  if( iQBits )
+    thres = TCoeff( ( int64_t( m_thrVal ) << ( iQBits - 1 ) ) );
+  else
+    thres = TCoeff( ( int64_t( m_thrVal >> 1 ) << iQBits ) );
+
+  useThres = thres / ( defaultQuantisationCoefficient << 2 );
+
+  const bool is4x4sbb = log2CGSize == 4 && cctx.log2CGWidth() == 2;
+
+  int subSetId = iScanPos >> log2CGSize;
+  for( ; subSetId >= 1; subSetId-- )
+  {
+    if( is4x4sbb && iScanPos >= 16 )
+    {
+      int  iScanPosinCG = iScanPos & ( iCGSize - 1 );
+      bool allSmaller   = true;
+
+      for( int xScanPosinCG = iScanPosinCG, xScanPos = iScanPos; allSmaller && xScanPosinCG >= 0; xScanPosinCG--, xScanPos-- )
+      {
+        const uint32_t uiBlkPos = cctx.blockPos( xScanPos );
+        allSmaller &= abs( piCoef.buf[uiBlkPos] ) <= useThres;
+      }
+
+      if( allSmaller )
+      {
+        iScanPos    -= iScanPosinCG + 1;
+        continue;
+      }
+      else
+      {
+        break;
+      }
+    }
+  }
+
   const int qBits8 = iQBits - 8;
   piQCoef.memset( 0 );
-  for (int uiBlockPos = 0; uiBlockPos < maxNumberOfCoeffs; uiBlockPos++ )
+
+  for( int currPos = 0; currPos <= iScanPos; currPos++ )
   {
+    const int uiBlockPos  = cctx.blockPos( currPos );
     const TCoeff iLevel   = piCoef.buf[uiBlockPos];
     const TCoeff iSign    = (iLevel < 0 ? -1: 1);
 
@@ -148,6 +240,7 @@ static void QuantCore(const CCoeffBuf&  piCoef,CoeffSigBuf piQCoef,TCoeff &uiAbs
     piQCoef.buf[uiBlockPos] = Clip3<TCoeff>( entropyCodingMinimum, entropyCodingMaximum, quantisedCoefficient );
   } // for n
 
+  lastScanPos = iScanPos;
 }
 
 static void DeQuantCore(const int maxX,const int maxY,const int scale,const TCoeffSig* const piQCoef,const size_t piQCfStride,TCoeff   *const piCoef,const int rightShift,const int inputMaximum,const TCoeff transformMaximum)
@@ -685,16 +778,19 @@ void Quant::quant(TransformUnit& tu, const ComponentID compID, const CCoeffBuf& 
     const int64_t iAdd = int64_t(tu.cs->slice->isIRAP() ? 171 : 85) << int64_t(iQBits - 9);
     const int qBits8 = iQBits - 8;
 
-    const uint32_t lfnstIdx = tu.cu->lfnstIdx;
-    const int maxNumberOfCoeffs = lfnstIdx > 0 ? ((( uiWidth == 4 && uiHeight == 4 ) || ( uiWidth == 8 && uiHeight == 8) ) ? 8 : 16) : piQCoef.area();
+    int lastScanPos = -1;
 
     if (!enableScalingLists)
-      xQuant (piCoef,piQCoef,uiAbsSum,deltaU,maxNumberOfCoeffs,defaultQuantisationCoefficient,iQBits,iAdd,entropyCodingMinimum,entropyCodingMaximum,cctx.signHiding() );
+      xQuant(tu,compID,piCoef,piQCoef,uiAbsSum,lastScanPos,deltaU,defaultQuantisationCoefficient,iQBits,iAdd,entropyCodingMinimum,entropyCodingMaximum,cctx.signHiding(),m_thrVal);
     else
     {
+      const uint32_t lfnstIdx = tu.cu->lfnstIdx;
+      const int maxNumberOfCoeffs = lfnstIdx > 0 ? ( ( ( uiWidth == 4 && uiHeight == 4 ) || ( uiWidth == 8 && uiHeight == 8 ) ) ? 8 : 16 ) : piQCoef.area();
+
       piQCoef.memset( 0 );
-      for (int uiBlockPos = 0; uiBlockPos < maxNumberOfCoeffs; uiBlockPos++ )
+      for (int uiScanPos = 0; uiScanPos < maxNumberOfCoeffs; uiScanPos++ )
       {
+        const int uiBlockPos  = cctx.blockPos( uiScanPos );
         const TCoeff iLevel   = piCoef.buf[uiBlockPos];
         const TCoeff iSign    = (iLevel < 0 ? -1: 1);
 
@@ -713,10 +809,9 @@ void Quant::quant(TransformUnit& tu, const ComponentID compID, const CCoeffBuf& 
       fwdResDPCM( tu, compID );
     }
 
-    int lastScanPos = -1;
     if( uiAbsSum )
     {
-      for( int scanPos = cctx.maxNumCoeff() - 1; scanPos >= 0; scanPos-- )
+      for( int scanPos = lastScanPos; scanPos >= 0; scanPos-- )
       {
         unsigned blkPos = cctx.blockPos( scanPos );
         if( piQCoef.buf[blkPos] )
