+++ source/Lib/CommonLib/x86/QuantX86.h
@@ -216,73 +216,170 @@ static void DeQuantCoreSIMD(const int maxX,const int maxY,const int scale,const 
 #define _mm_storeu_si32(p, a) (void)(*(int*)(p) = _mm_cvtsi128_si32((a)))
 
 template<X86_VEXT vext>
-static void QuantCoreSIMD(const CCoeffBuf&  piCoef,CoeffSigBuf piQCoef,TCoeff &uiAbsSum,TCoeff *deltaU,const int maxNumberOfCoeffs,const int defaultQuantisationCoefficient,const int iQBits,const int64_t iAdd,const TCoeff entropyCodingMinimum,const TCoeff entropyCodingMaximum ,const bool signHiding)
+static void QuantCoreSIMD(const TransformUnit tu, const ComponentID compID, const CCoeffBuf& piCoef,CoeffSigBuf piQCoef,TCoeff &uiAbsSum, int &lastScanPos,TCoeff *deltaU,const int defaultQuantisationCoefficient,const int iQBits,const int64_t iAdd,const TCoeff entropyCodingMinimum,const TCoeff entropyCodingMaximum,const bool signHiding, const TCoeff m_thrVal)
 {
+  CoeffCodingContext cctx( tu, compID, signHiding );
+
+  const CompArea &rect      = tu.blocks[compID];
+  const uint32_t uiWidth    = rect.width;
+  const uint32_t uiHeight   = rect.height;
+  const uint32_t log2CGSize = cctx.log2CGSize();
+
+  uiAbsSum = 0;
+
+  const int iCGSize   = 1 << log2CGSize;
+
+  const uint32_t lfnstIdx = tu.cu->lfnstIdx;
+  const int iCGNum   = lfnstIdx > 0 ? 1 : std::min<int>(JVET_C0024_ZERO_OUT_TH, uiWidth) * std::min<int>(JVET_C0024_ZERO_OUT_TH, uiHeight) >> cctx.log2CGSize();
+  int       iScanPos = ( iCGNum << log2CGSize ) - 1;
+
+  if( lfnstIdx > 0 && ( ( uiWidth == 4 && uiHeight == 4 ) || ( uiWidth == 8 && uiHeight == 8 ) ) )
+  {
+    iScanPos = 7;
+  }
+
+  // Find first non-zero coeff
+  for( ; iScanPos > 0; iScanPos-- )
+  {
+    uint32_t uiBlkPos = cctx.blockPos( iScanPos );
+    if( piCoef.buf[uiBlkPos] )
+      break;
+  }
+
+  //////////////////////////////////////////////////////////////////////////
+  //  Loop over sub-sets (coefficient groups)
+  //////////////////////////////////////////////////////////////////////////
+  
+  TCoeff thres = 0, useThres = 0;
+  
+  if( iQBits )
+    thres = TCoeff( ( int64_t( m_thrVal ) << ( iQBits - 1 ) ) );
+  else
+    thres = TCoeff( ( int64_t( m_thrVal >> 1 ) << iQBits ) );
+
+  useThres = thres / ( defaultQuantisationCoefficient << 2 );
+
+  const bool is4x4sbb = log2CGSize == 4 && cctx.log2CGWidth() == 2;
+
+  int subSetId = iScanPos >> log2CGSize;
+  // if more than one 4x4 coding subblock is available, use SIMD to find first subblock with coefficient larger than threshold
+  if( is4x4sbb && iScanPos >= 16 )
+  {
+    for( ; subSetId >= 1; subSetId-- )
+    {
+      // move the pointer to the beginning of the current subblock
+      const int iScanPosinCG = iScanPos & ( iCGSize - 1 );
+      const int firstTestPos = iScanPos - iScanPosinCG;
+      uint32_t  uiBlkPos     = cctx.blockPos( firstTestPos );
+
+      const __m128i xdfTh = _mm_set1_epi32( useThres );
+
+      // read first line of the subblock and check for coefficients larger than the threshold
+      // assumming the subblocks are dense 4x4 blocks in raster scan order with the stride of tuPars.m_width
+      __m128i xl0 = _mm_abs_epi32( _mm_loadu_si128( ( const __m128i* ) &piCoef.buf[uiBlkPos] ) );
+      __m128i xdf = _mm_cmpgt_epi32( xl0, xdfTh );
+
+      // same for the next line in the subblock
+      uiBlkPos += uiWidth;
+      xl0 = _mm_abs_epi32( _mm_loadu_si128( ( const __m128i* ) &piCoef.buf[uiBlkPos] ) );
+      xdf = _mm_or_si128( xdf, _mm_cmpgt_epi32( xl0, xdfTh ) );
+
+      // and the third line
+      uiBlkPos += uiWidth;
+      xl0 = _mm_abs_epi32( _mm_loadu_si128( ( const __m128i* ) &piCoef.buf[uiBlkPos] ) );
+      xdf = _mm_or_si128( xdf, _mm_cmpgt_epi32( xl0, xdfTh ) );
+
+      // and the last line
+      uiBlkPos += uiWidth;
+      xl0 = _mm_abs_epi32( _mm_loadu_si128( ( const __m128i* ) &piCoef.buf[uiBlkPos] ) );
+      xdf = _mm_or_si128( xdf, _mm_cmpgt_epi32( xl0, xdfTh ) );
+
+      if( _mm_testz_si128( xdf, xdf ) )
+      {
+        iScanPos -= iScanPosinCG + 1;
+        continue;
+      }
+      else
+      {
+        break;
+      }
+    }
+  }
+
   const int qBits8 = iQBits - 8;
   piQCoef.memset( 0 );
 
-  if( maxNumberOfCoeffs >= 8 )
+  lastScanPos = iScanPos;
+
+  if( is4x4sbb && ( iScanPos & 15 ) == 15 )
   {
+#if defined( USE_AVX2 ) && 0 // sometimes has undefined behavior
     if( vext >= AVX2 )
     {
-#ifdef USE_AVX2
       const __m256i vNull   = _mm256_setzero_si256();
       const __m256i vQuantCoeff = _mm256_set1_epi32(defaultQuantisationCoefficient);
-      const __m256i vAdd    = _mm256_set_epi64x(iAdd,iAdd,iAdd,iAdd);
+      const __m256i vAdd    = _mm256_set1_epi64x(iAdd);
       const __m256i vMax    = _mm256_set1_epi32(entropyCodingMaximum);
       const __m256i vMin    = _mm256_set1_epi32(entropyCodingMinimum);
       const __m256i vMask   = _mm256_set_epi32( 0, -1, 0, -1, 0, -1, 0, -1 );
       __m256i vAbsSum = vNull;
 
-      for (int uiBlockPos = 0; uiBlockPos < maxNumberOfCoeffs; uiBlockPos+=8 )
+      for( subSetId = iScanPos >> log2CGSize; subSetId >= 0; subSetId-- )
       {
-        __m256i  vLevel = _mm256_loadu_si256((__m256i *)&piCoef.buf[uiBlockPos]);     // coeff7,coeff6,coeff5,coeff4,coeff3,coeff2,coeff1,coeff0,
-        __m256i vSign = _mm256_cmpgt_epi32 (vNull,vLevel);                                            // sign3,sign2,sign1,sign0 FFFF or 0000
-        vLevel = _mm256_abs_epi32 (vLevel);
-        __m256i vdeltaU0 = _mm256_mul_epu32(vLevel,vQuantCoeff);                      // Tmp2,Tmp0
-        __m256i  vdeltaU1 = _mm256_srli_si256(vLevel,4);                                            // abs(0,vLevel3,vLevel2,vLevel1)
-        vdeltaU1 = _mm256_mul_epu32(vdeltaU1,vQuantCoeff);                          // Tmp3,Tmp1
-        __m256i vTmpLevel_0 = _mm256_add_epi64(vdeltaU0,vAdd);
-        __m256i vTmpLevel_1 = _mm256_add_epi64(vdeltaU1,vAdd);
-        vTmpLevel_0 = _mm256_srli_epi64(vTmpLevel_0,iQBits);                                         // Int32 Tmp2,Tmp0
-        vTmpLevel_1 = _mm256_srli_epi64(vTmpLevel_1,iQBits);                                         // Int32 Tmp3,Tmp1
+        int uiBlockPos = cctx.blockPos( subSetId << log2CGSize );
 
-        if (signHiding)
+        for( int line = 0; line < 4; line += 2, uiBlockPos += ( 2 * uiWidth ) )
         {
-          __m256i vBS0 = _mm256_slli_epi64(vTmpLevel_0,iQBits);
-          __m256i vBS1 = _mm256_slli_epi64(vTmpLevel_1,iQBits);
-          vdeltaU0 = _mm256_sub_epi64(vdeltaU0,vBS0);
-          vdeltaU1 = _mm256_sub_epi64(vdeltaU1,vBS1);
-          vdeltaU0 = _mm256_srli_epi64(vdeltaU0,qBits8);
-          vdeltaU1 = _mm256_srli_epi64(vdeltaU1,qBits8);
-          vdeltaU0 = _mm256_and_si256(vdeltaU0,vMask);
-          vdeltaU1 = _mm256_and_si256(vdeltaU1,vMask);
-          vdeltaU1 = _mm256_slli_epi64(vdeltaU1,32);
-          vdeltaU0 = _mm256_or_si256(vdeltaU0,vdeltaU1);
-          _mm256_storeu_si256( ( __m256i * )&deltaU[uiBlockPos],vdeltaU0);
+          __m256i  vLevel = _mm256_castsi128_si256 (         _mm_loadu_si128((__m128i *)&piCoef.buf[uiBlockPos]           )    ); // coeff7,coeff6,coeff5,coeff4,coeff3,coeff2,coeff1,coeff0,
+                   vLevel = _mm256_inserti128_si256( vLevel, _mm_loadu_si128((__m128i *)&piCoef.buf[uiBlockPos + uiWidth] ), 1 ); // coeff7,coeff6,coeff5,coeff4,coeff3,coeff2,coeff1,coeff0,
+          __m256i vSign = _mm256_cmpgt_epi32 (vNull,vLevel);                            // sign3,sign2,sign1,sign0 FFFF or 0000
+          vLevel = _mm256_abs_epi32 (vLevel);
+          __m256i vdeltaU0 = _mm256_mul_epu32(vLevel,vQuantCoeff);                      // Tmp2,Tmp0
+          __m256i  vdeltaU1 = _mm256_srli_si256(vLevel,4);                              // abs(0,vLevel3,vLevel2,vLevel1)
+          vdeltaU1 = _mm256_mul_epu32(vdeltaU1,vQuantCoeff);                            // Tmp3,Tmp1
+          __m256i vTmpLevel_0 = _mm256_add_epi64(vdeltaU0,vAdd);
+          __m256i vTmpLevel_1 = _mm256_add_epi64(vdeltaU1,vAdd);
+          vTmpLevel_0 = _mm256_srli_epi64(vTmpLevel_0,iQBits);                          // Int32 Tmp2,Tmp0
+          vTmpLevel_1 = _mm256_srli_epi64(vTmpLevel_1,iQBits);                          // Int32 Tmp3,Tmp1
+
+          if (signHiding)
+          {
+            __m256i vBS0 = _mm256_slli_epi64(vTmpLevel_0,iQBits);
+            __m256i vBS1 = _mm256_slli_epi64(vTmpLevel_1,iQBits);
+            vdeltaU0 = _mm256_sub_epi64(vdeltaU0,vBS0);
+            vdeltaU1 = _mm256_sub_epi64(vdeltaU1,vBS1);
+            vdeltaU0 = _mm256_srli_epi64(vdeltaU0,qBits8);
+            vdeltaU1 = _mm256_srli_epi64(vdeltaU1,qBits8);
+            vdeltaU0 = _mm256_and_si256(vdeltaU0,vMask);
+            vdeltaU1 = _mm256_and_si256(vdeltaU1,vMask);
+            vdeltaU1 = _mm256_slli_epi64(vdeltaU1,32);
+            vdeltaU0 = _mm256_or_si256(vdeltaU0,vdeltaU1);
+            _mm_storeu_si128( ( __m128i * )&deltaU[uiBlockPos],          _mm256_castsi256_si128  (vdeltaU0));
+            _mm_storeu_si128( ( __m128i * )&deltaU[uiBlockPos + uiWidth],_mm256_extracti128_si256(vdeltaU0, 1));
+          }
+          __m256i vquantMag0 = _mm256_and_si256(vTmpLevel_0,vMask);
+          __m256i vquantMag1 = _mm256_and_si256(vTmpLevel_1,vMask);
+          vquantMag1  = _mm256_slli_epi64(vquantMag1,32);
+          vTmpLevel_0 = _mm256_or_si256(vquantMag0,vquantMag1);
+          vAbsSum     = _mm256_add_epi32(vAbsSum,vTmpLevel_0);
+          vTmpLevel_1 = _mm256_and_si256(vTmpLevel_0,vSign);                            // mask only neg values
+          vTmpLevel_0 = _mm256_andnot_si256(vSign,vTmpLevel_0);                         // mask only pos values
+          vTmpLevel_0 = _mm256_sub_epi32(vTmpLevel_0,vTmpLevel_1);
+          vTmpLevel_0 = _mm256_min_epi32(vMax, _mm256_max_epi32(vMin,vTmpLevel_0));     // clip to 16 Bit
+          vTmpLevel_0 = _mm256_packs_epi32(vTmpLevel_0,vTmpLevel_0);
+          _mm_storel_epi64( ( __m128i * )&piQCoef.buf[uiBlockPos],          _mm256_castsi256_si128  (vTmpLevel_0));
+          _mm_storel_epi64( ( __m128i * )&piQCoef.buf[uiBlockPos + uiWidth],_mm256_extracti128_si256(vTmpLevel_0, 1));
         }
-        __m256i vquantMag0 = _mm256_and_si256(vTmpLevel_0,vMask);
-        __m256i vquantMag1 = _mm256_and_si256(vTmpLevel_1,vMask);
-        vquantMag1  = _mm256_slli_epi64(vquantMag1,32);
-        vTmpLevel_0 = _mm256_or_si256(vquantMag0,vquantMag1);
-        vAbsSum     = _mm256_add_epi32(vAbsSum,vTmpLevel_0);
-        vTmpLevel_1 = _mm256_and_si256(vTmpLevel_0,vSign);                                       // mask only neg values
-        vTmpLevel_0 = _mm256_andnot_si256(vSign,vTmpLevel_0);                                 // mask only pos values
-        vTmpLevel_0 = _mm256_sub_epi32(vTmpLevel_0,vTmpLevel_1);
-        vTmpLevel_0 = _mm256_min_epi32(vMax, _mm256_max_epi32(vMin,vTmpLevel_0));  // clip to 16 Bit
-        vTmpLevel_0 = _mm256_packs_epi32(vTmpLevel_0,vTmpLevel_0);
-        vTmpLevel_0 = _mm256_permute4x64_epi64(vTmpLevel_0, ( 0 << 0 ) + ( 2 << 2 ) );
-        _mm_storeu_si128( ( __m128i * )&piQCoef.buf[uiBlockPos],_mm256_castsi256_si128(vTmpLevel_0));
       }
 
       __m128i xAbsSum = _mm_add_epi32( _mm256_castsi256_si128( vAbsSum ), _mm256_extracti128_si256( vAbsSum, 1 ) );
       xAbsSum = _mm_hadd_epi32( xAbsSum, xAbsSum );
       xAbsSum = _mm_hadd_epi32( xAbsSum, xAbsSum );
 
-      uiAbsSum = _mm_cvtsi128_si32( xAbsSum );
-#endif
+      uiAbsSum += _mm_cvtsi128_si32( xAbsSum );
     }  //AVX2
     else
+#endif
     {
       const __m128i vNull = _mm_setzero_si128();
       const __m128i vQuantCoeff = _mm_set1_epi32(defaultQuantisationCoefficient);
@@ -292,55 +389,61 @@ static void QuantCoreSIMD(const CCoeffBuf&  piCoef,CoeffSigBuf piQCoef,TCoeff &u
       const __m128i vMask = _mm_set_epi32(0, -1, 0, -1);
       __m128i vAbsSum = vNull;
 
-      for (int uiBlockPos = 0; uiBlockPos < maxNumberOfCoeffs; uiBlockPos+=4 )
+      for( subSetId = iScanPos >> log2CGSize; subSetId >= 0; subSetId-- )
       {
-        __m128i vLevel = _mm_lddqu_si128((__m128i*)&piCoef.buf[uiBlockPos]);      // coeff3,coeff2,coeff1,coeff0,
-        __m128i vSign = _mm_cmpgt_epi32 (vNull,vLevel);                                            // sign3,sign2,sign1,sign0 FFFF or 0000
-        vLevel = _mm_abs_epi32 (vLevel);                                                                         // abs(vLevel3,vLevel2,vLevel1,vLevel0)
-        __m128i vdeltaU0 = _mm_mul_epu32(vLevel,vQuantCoeff);                             // Tmp2,Tmp0
-        __m128i vdeltaU1 = _mm_bsrli_si128(vLevel,4);                                                  // abs(0,vLevel3,vLevel2,vLevel1)
-        vdeltaU1 = _mm_mul_epu32(vdeltaU1,vQuantCoeff);                                       // Tmp3,Tmp1
-        __m128i vTmpLevel_0 = _mm_add_epi64(vdeltaU0,vAdd);
-        __m128i vTmpLevel_1 = _mm_add_epi64(vdeltaU1,vAdd);
-        vTmpLevel_0 = _mm_srli_epi64(vTmpLevel_0,iQBits);                                         // Int32 Tmp2,Tmp0
-        vTmpLevel_1 = _mm_srli_epi64(vTmpLevel_1,iQBits);                                         // Int32 Tmp3,Tmp1
-        if (signHiding)
+        int uiBlockPos = cctx.blockPos( subSetId << log2CGSize );
+
+        for( int line = 0; line < 4; line++, uiBlockPos += uiWidth )
         {
-          __m128i vBS0 = _mm_slli_epi64(vTmpLevel_0,iQBits);
-          __m128i vBS1 = _mm_slli_epi64(vTmpLevel_1,iQBits);
-          vdeltaU0 = _mm_sub_epi64(vdeltaU0,vBS0);
-          vdeltaU1 = _mm_sub_epi64(vdeltaU1,vBS1);
-          vdeltaU0 = _mm_srli_epi64(vdeltaU0,qBits8);
-          vdeltaU1 = _mm_srli_epi64(vdeltaU1,qBits8);
-          vdeltaU0 =  _mm_and_si128(vdeltaU0,vMask);
-          vdeltaU1 =  _mm_and_si128(vdeltaU1,vMask);
-          vdeltaU1 =   _mm_slli_epi64(vdeltaU1,32);
-          vdeltaU0 = _mm_or_si128(vdeltaU0,vdeltaU1);
-         _mm_storeu_si128( ( __m128i * )&deltaU[uiBlockPos],vdeltaU0);
+          __m128i vLevel = _mm_loadu_si128((__m128i*)&piCoef.buf[uiBlockPos]);      // coeff3,coeff2,coeff1,coeff0,
+          __m128i vSign = _mm_cmpgt_epi32 (vNull,vLevel);                           // sign3,sign2,sign1,sign0 FFFF or 0000
+          vLevel = _mm_abs_epi32 (vLevel);                                          // abs(vLevel3,vLevel2,vLevel1,vLevel0)
+          __m128i vdeltaU0 = _mm_mul_epu32(vLevel,vQuantCoeff);                     // Tmp2,Tmp0
+          __m128i vdeltaU1 = _mm_srli_si128(vLevel,4);                             // abs(0,vLevel3,vLevel2,vLevel1)
+          vdeltaU1 = _mm_mul_epu32(vdeltaU1,vQuantCoeff);                           // Tmp3,Tmp1
+          __m128i vTmpLevel_0 = _mm_add_epi64(vdeltaU0,vAdd);
+          __m128i vTmpLevel_1 = _mm_add_epi64(vdeltaU1,vAdd);
+          vTmpLevel_0 = _mm_srli_epi64(vTmpLevel_0,iQBits);                         // Int32 Tmp2,Tmp0
+          vTmpLevel_1 = _mm_srli_epi64(vTmpLevel_1,iQBits);                         // Int32 Tmp3,Tmp1
+          if (signHiding)
+          {
+            __m128i vBS0 = _mm_slli_epi64(vTmpLevel_0,iQBits);
+            __m128i vBS1 = _mm_slli_epi64(vTmpLevel_1,iQBits);
+            vdeltaU0 = _mm_sub_epi64(vdeltaU0,vBS0);
+            vdeltaU1 = _mm_sub_epi64(vdeltaU1,vBS1);
+            vdeltaU0 = _mm_srli_epi64(vdeltaU0,qBits8);
+            vdeltaU1 = _mm_srli_epi64(vdeltaU1,qBits8);
+            vdeltaU0 = _mm_and_si128(vdeltaU0,vMask);
+            vdeltaU1 = _mm_and_si128(vdeltaU1,vMask);
+            vdeltaU1 = _mm_slli_epi64(vdeltaU1,32);
+            vdeltaU0 = _mm_or_si128(vdeltaU0,vdeltaU1);
+           _mm_storeu_si128( ( __m128i * ) &deltaU[uiBlockPos],vdeltaU0);
+          }
+          __m128i vquantMag0 =  _mm_and_si128(vTmpLevel_0,vMask);
+          __m128i vquantMag1 =  _mm_and_si128(vTmpLevel_1,vMask);
+          vquantMag1  = _mm_slli_epi64(vquantMag1,32);
+          vTmpLevel_0 = _mm_or_si128(vquantMag0,vquantMag1);
+          vAbsSum     = _mm_add_epi32(vAbsSum,vTmpLevel_0);
+          vTmpLevel_1 = _mm_and_si128(vTmpLevel_0,vSign);                           // mask only neg values
+          vTmpLevel_0 = _mm_andnot_si128(vSign,vTmpLevel_0);                        // mask only pos values
+          vTmpLevel_0 = _mm_sub_epi32(vTmpLevel_0,vTmpLevel_1);
+          vTmpLevel_0 = _mm_min_epi32(vMax, _mm_max_epi32(vMin,vTmpLevel_0));       // clip to 16 Bit
+          vTmpLevel_0 = _mm_packs_epi32(vTmpLevel_0,vTmpLevel_0);
+          _mm_storel_epi64( ( __m128i * ) &piQCoef.buf[uiBlockPos],vTmpLevel_0);
         }
-        __m128i vquantMag0 =  _mm_and_si128(vTmpLevel_0,vMask);
-        __m128i vquantMag1 =  _mm_and_si128(vTmpLevel_1,vMask);
-        vquantMag1 = _mm_slli_epi64(vquantMag1,32);
-        vTmpLevel_0 = _mm_or_si128(vquantMag0,vquantMag1);
-        vAbsSum = _mm_add_epi32(vAbsSum,vTmpLevel_0);
-        vTmpLevel_1 = _mm_and_si128(vTmpLevel_0,vSign);                                       // mask only neg values
-        vTmpLevel_0 = _mm_andnot_si128(vSign,vTmpLevel_0);                                 // mask only pos values
-        vTmpLevel_0 = _mm_sub_epi32(vTmpLevel_0,vTmpLevel_1);
-        vTmpLevel_0 = _mm_min_epi32(vMax, _mm_max_epi32(vMin,vTmpLevel_0));  // clip to 16 Bit
-        vTmpLevel_0 = _mm_packs_epi32(vTmpLevel_0,vTmpLevel_0);
-        _mm_storel_epi64( ( __m128i * )&piQCoef.buf[uiBlockPos],vTmpLevel_0);
       }
 
       vAbsSum = _mm_hadd_epi32( vAbsSum, vAbsSum );
       vAbsSum = _mm_hadd_epi32( vAbsSum, vAbsSum );
 
-      uiAbsSum = _mm_cvtsi128_si32( vAbsSum );
+      uiAbsSum += _mm_cvtsi128_si32( vAbsSum );
     }
   }
   else
   {
-    for (int uiBlockPos = 0; uiBlockPos < maxNumberOfCoeffs; uiBlockPos++ )
+    for( int currPos = 0; currPos <= iScanPos; currPos++ )
     {
+      const int uiBlockPos  = cctx.blockPos( currPos );
       const TCoeff iLevel   = piCoef.buf[uiBlockPos];
       const TCoeff iSign    = (iLevel < 0 ? -1: 1);
       const int64_t  tmpLevel = (int64_t)abs(iLevel) * defaultQuantisationCoefficient;
